Добавление фукнционала удаления всего списка + декомпозиция

+++ src/App.jsx
@@ -1,5 +1,8 @@
 import { useEffect, useState } from "react";
 import "./index.css";
+import { Form } from "./components/Form";
+import { Itemlist } from "./components/Itemlist";
+import { Stats } from "./components/Stats";
 
 const DEFAULT_ITEMS = [];
 const STORAGE_KEY = "todo-items";
@@ -38,7 +41,9 @@ export function App() {
   }
 
   function handleDeleteAllItem() {
-    setItem([]);
+    const confirmed = window.confirm("Вы уверены, что хотите все удалить?");
+
+    if (confirmed) setItem([]);
   }
 
   return (
@@ -52,123 +57,6 @@ export function App() {
         onDeleteAllItem={handleDeleteAllItem}
       />
       <Stats item={item} />
-    </div>
-  );
-}
-
-function Form({ onAddItems }) {
-  const [active, setActive] = useState("");
-
-  function handleSubmit(e) {
-    e.preventDefault();
-
-    if (!active) return;
-
-    const newItem = { active, completed: false, id: Date.now() };
-    console.log(newItem);
-
-    onAddItems(newItem);
-
-    setActive("");
-  }
-
-  return (
-    <form className="todo-form" onSubmit={handleSubmit}>
-      <input
-        className="todo-input"
-        type="text"
-        placeholder="Новая задача"
-        value={active}
-        onChange={(e) => setActive(e.target.value)}
-      />
-      <button className="todo-add-btn">+</button>
-    </form>
-  );
-}
-
-function Itemlist({ item, onDeleteItem, onToggleItem, onDeleteAllItem }) {
-  const [sortBy, setSortBy] = useState("input");
-
-  let sortedItems;
-  if (sortBy === "input") sortedItems = item;
-
-  if (sortBy === "active")
-    sortedItems = item.slice().sort((a, b) => a.active.localeCompare(b.active));
-
-  if (sortBy === "completed")
-    sortedItems = item
-      .slice()
-      .sort((a, b) => Number(a.completed) - Number(b.completed));
-
-  if (sortedItems.length === 0) return <></>;
-  if (sortedItems.length > 0)
-    return (
-      <>
-        <ul className="todo-list">
-          {sortedItems.map((item) => (
-            <Item
-              item={item}
-              key={item.id}
-              onDeleteItem={onDeleteItem}
-              onToggleItem={onToggleItem}
-            />
-          ))}
-        </ul>
-        <div className="todo-controls">
-          <select
-            value={sortBy}
-            onChange={(e) => setSortBy(e.target.value)}
-            className="todo-select"
-          >
-            <option value="input">Сортировка по добавлению</option>
-            <option value="active">Сортировка по алфавиту</option>
-            <option value="completed">Сотировка по выполнению</option>
-          </select>
-          <button className="todo-reset" onClick={onDeleteAllItem}>
-            Очистить поле
-          </button>
-        </div>
-      </>
-    );
-}
-
-function Item({ item, onDeleteItem, onToggleItem }) {
-  return (
-    <li className="todo-item">
-      <label className="todo-check">
-        <input type="checkbox" onChange={() => onToggleItem(item.id)}></input>
-        <span
-          className="todo-text"
-          style={item.completed ? { textDecoration: "line-through" } : {}}
-        >
-          {item.active}
-        </span>
-      </label>
-      <button className="todo-delete" onClick={() => onDeleteItem(item.id)}>
-        Х
-      </button>
-    </li>
-  );
-}
-
-function Stats({ item }) {
-  if (!item.length)
-    return (
-      <p className="footer">
-        <em>Начни добавлять задачи</em>
-      </p>
-    );
-
-  const numItems = item.length;
-  const numCompleted = item.filter((item) => item.completed).length;
-  const percentage = Math.round((numCompleted / numItems) * 100);
-  return (
-    <div className="footer">
-      <em>
-        {percentage === 100
-          ? "Ты молодец, все сделано"
-          : `Задач всего: ${numItems}. Сделано: ${numCompleted} (${percentage}%)`}
-      </em>
     </div>
   );
 }
